+++ src/chunk.h
@@ -15,10 +15,10 @@ class Chunk {
   public:
     Chunk();
 
-    void write_constant(value val, u32 line);
+    void write_constant(value val, u64 line);
 
-    void write(const u8& byte, u32 line);
-    void write(const OpCode& op, u32 line);
+    void write(const u8& byte, u64 line);
+    void write(const OpCode& op, u64 line);
 
     [[nodiscard]] size_t add_constant(value val);
